requirements.txt
Django==1.9.7
django-extensions==1.6.7
django-filter
djangorestframework
Markdown
